working update jul 31: working on old-laptop-server>page.mdx
