+++ .open-next/cloudflare/cache-assets-manifest.sql
@@ -1,3 +1,3 @@
 CREATE TABLE IF NOT EXISTS tags (tag TEXT NOT NULL, path TEXT NOT NULL, UNIQUE(tag, path) ON CONFLICT REPLACE);
      CREATE TABLE IF NOT EXISTS revalidations (tag TEXT NOT NULL, revalidatedAt INTEGER NOT NULL, UNIQUE(tag) ON CONFLICT REPLACE);
-INSERT INTO tags (tag, path) VALUES ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/layout", "MI4X1be_6PxCG5Hjbsi0O/_not-found"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/_not-found/layout", "MI4X1be_6PxCG5Hjbsi0O/_not-found"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/_not-found/page", "MI4X1be_6PxCG5Hjbsi0O/_not-found"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/_not-found", "MI4X1be_6PxCG5Hjbsi0O/_not-found"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/layout", "MI4X1be_6PxCG5Hjbsi0O/blog/example-mdx-metadata"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/blog/layout", "MI4X1be_6PxCG5Hjbsi0O/blog/example-mdx-metadata"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/blog/example-mdx-metadata/layout", "MI4X1be_6PxCG5Hjbsi0O/blog/example-mdx-metadata"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/blog/example-mdx-metadata/page", "MI4X1be_6PxCG5Hjbsi0O/blog/example-mdx-metadata"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/blog/example-mdx-metadata", "MI4X1be_6PxCG5Hjbsi0O/blog/example-mdx-metadata"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/layout", "MI4X1be_6PxCG5Hjbsi0O/blog/setting-up-cloudflare"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/blog/layout", "MI4X1be_6PxCG5Hjbsi0O/blog/setting-up-cloudflare"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/blog/setting-up-cloudflare/layout", "MI4X1be_6PxCG5Hjbsi0O/blog/setting-up-cloudflare"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/blog/setting-up-cloudflare/page", "MI4X1be_6PxCG5Hjbsi0O/blog/setting-up-cloudflare"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/blog/setting-up-cloudflare", "MI4X1be_6PxCG5Hjbsi0O/blog/setting-up-cloudflare"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/layout", "MI4X1be_6PxCG5Hjbsi0O/blog/setting-up-smtp"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/blog/layout", "MI4X1be_6PxCG5Hjbsi0O/blog/setting-up-smtp"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/blog/setting-up-smtp/layout", "MI4X1be_6PxCG5Hjbsi0O/blog/setting-up-smtp"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/blog/setting-up-smtp/page", "MI4X1be_6PxCG5Hjbsi0O/blog/setting-up-smtp"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/blog/setting-up-smtp", "MI4X1be_6PxCG5Hjbsi0O/blog/setting-up-smtp"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/layout", "MI4X1be_6PxCG5Hjbsi0O/favicon.ico"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/favicon.ico/layout", "MI4X1be_6PxCG5Hjbsi0O/favicon.ico"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/favicon.ico/route", "MI4X1be_6PxCG5Hjbsi0O/favicon.ico"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/favicon.ico", "MI4X1be_6PxCG5Hjbsi0O/favicon.ico"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/layout", "MI4X1be_6PxCG5Hjbsi0O/index"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/page", "MI4X1be_6PxCG5Hjbsi0O/index"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/", "MI4X1be_6PxCG5Hjbsi0O/index"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/layout", "MI4X1be_6PxCG5Hjbsi0O/projects/exploring-virt"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/projects/layout", "MI4X1be_6PxCG5Hjbsi0O/projects/exploring-virt"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/projects/exploring-virt/layout", "MI4X1be_6PxCG5Hjbsi0O/projects/exploring-virt"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/projects/exploring-virt/page", "MI4X1be_6PxCG5Hjbsi0O/projects/exploring-virt"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/projects/exploring-virt", "MI4X1be_6PxCG5Hjbsi0O/projects/exploring-virt"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/layout", "MI4X1be_6PxCG5Hjbsi0O/projects/old-laptop-server"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/projects/layout", "MI4X1be_6PxCG5Hjbsi0O/projects/old-laptop-server"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/projects/old-laptop-server/layout", "MI4X1be_6PxCG5Hjbsi0O/projects/old-laptop-server"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/projects/old-laptop-server/page", "MI4X1be_6PxCG5Hjbsi0O/projects/old-laptop-server"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/projects/old-laptop-server", "MI4X1be_6PxCG5Hjbsi0O/projects/old-laptop-server"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/layout", "MI4X1be_6PxCG5Hjbsi0O/robots.txt"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/robots.txt/layout", "MI4X1be_6PxCG5Hjbsi0O/robots.txt"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/robots.txt/route", "MI4X1be_6PxCG5Hjbsi0O/robots.txt"), ("MI4X1be_6PxCG5Hjbsi0O/_N_T_/robots.txt", "MI4X1be_6PxCG5Hjbsi0O/robots.txt");
+INSERT INTO tags (tag, path) VALUES ("zQgnelpe2z0ts-dbnrLzn/_N_T_/layout", "zQgnelpe2z0ts-dbnrLzn/_not-found"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/_not-found/layout", "zQgnelpe2z0ts-dbnrLzn/_not-found"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/_not-found/page", "zQgnelpe2z0ts-dbnrLzn/_not-found"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/_not-found", "zQgnelpe2z0ts-dbnrLzn/_not-found"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/layout", "zQgnelpe2z0ts-dbnrLzn/blog/example-mdx-metadata"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/blog/layout", "zQgnelpe2z0ts-dbnrLzn/blog/example-mdx-metadata"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/blog/example-mdx-metadata/layout", "zQgnelpe2z0ts-dbnrLzn/blog/example-mdx-metadata"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/blog/example-mdx-metadata/page", "zQgnelpe2z0ts-dbnrLzn/blog/example-mdx-metadata"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/blog/example-mdx-metadata", "zQgnelpe2z0ts-dbnrLzn/blog/example-mdx-metadata"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/layout", "zQgnelpe2z0ts-dbnrLzn/blog/setting-up-cloudflare"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/blog/layout", "zQgnelpe2z0ts-dbnrLzn/blog/setting-up-cloudflare"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/blog/setting-up-cloudflare/layout", "zQgnelpe2z0ts-dbnrLzn/blog/setting-up-cloudflare"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/blog/setting-up-cloudflare/page", "zQgnelpe2z0ts-dbnrLzn/blog/setting-up-cloudflare"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/blog/setting-up-cloudflare", "zQgnelpe2z0ts-dbnrLzn/blog/setting-up-cloudflare"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/layout", "zQgnelpe2z0ts-dbnrLzn/blog/setting-up-smtp"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/blog/layout", "zQgnelpe2z0ts-dbnrLzn/blog/setting-up-smtp"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/blog/setting-up-smtp/layout", "zQgnelpe2z0ts-dbnrLzn/blog/setting-up-smtp"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/blog/setting-up-smtp/page", "zQgnelpe2z0ts-dbnrLzn/blog/setting-up-smtp"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/blog/setting-up-smtp", "zQgnelpe2z0ts-dbnrLzn/blog/setting-up-smtp"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/layout", "zQgnelpe2z0ts-dbnrLzn/favicon.ico"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/favicon.ico/layout", "zQgnelpe2z0ts-dbnrLzn/favicon.ico"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/favicon.ico/route", "zQgnelpe2z0ts-dbnrLzn/favicon.ico"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/favicon.ico", "zQgnelpe2z0ts-dbnrLzn/favicon.ico"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/layout", "zQgnelpe2z0ts-dbnrLzn/index"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/page", "zQgnelpe2z0ts-dbnrLzn/index"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/", "zQgnelpe2z0ts-dbnrLzn/index"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/layout", "zQgnelpe2z0ts-dbnrLzn/projects/exploring-virt"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/projects/layout", "zQgnelpe2z0ts-dbnrLzn/projects/exploring-virt"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/projects/exploring-virt/layout", "zQgnelpe2z0ts-dbnrLzn/projects/exploring-virt"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/projects/exploring-virt/page", "zQgnelpe2z0ts-dbnrLzn/projects/exploring-virt"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/projects/exploring-virt", "zQgnelpe2z0ts-dbnrLzn/projects/exploring-virt"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/layout", "zQgnelpe2z0ts-dbnrLzn/projects/old-laptop-server"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/projects/layout", "zQgnelpe2z0ts-dbnrLzn/projects/old-laptop-server"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/projects/old-laptop-server/layout", "zQgnelpe2z0ts-dbnrLzn/projects/old-laptop-server"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/projects/old-laptop-server/page", "zQgnelpe2z0ts-dbnrLzn/projects/old-laptop-server"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/projects/old-laptop-server", "zQgnelpe2z0ts-dbnrLzn/projects/old-laptop-server"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/layout", "zQgnelpe2z0ts-dbnrLzn/robots.txt"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/robots.txt/layout", "zQgnelpe2z0ts-dbnrLzn/robots.txt"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/robots.txt/route", "zQgnelpe2z0ts-dbnrLzn/robots.txt"), ("zQgnelpe2z0ts-dbnrLzn/_N_T_/robots.txt", "zQgnelpe2z0ts-dbnrLzn/robots.txt");
