+++ .wrangler/state/v3/r2/walkable-blog-cache/blobs/0c4528474f00842528645a0e827083eea2cf8576d1fedbba9769a6350b6d9b1e000001985fbb795c
@@ -1,0 +1,1 @@
+{"type":"app","html":"<!DOCTYPE html><html><head><meta charSet=\"utf-8\" data-next-head=\"\"/><meta name=\"viewport\" content=\"width=device-width\" data-next-head=\"\"/><title data-next-head=\"\">500: Internal Server Error</title><noscript data-n-css=\"\"></noscript><script defer=\"\" noModule=\"\" src=\"/_next/static/chunks/polyfills-42372ed130431b0a.js\"></script><script src=\"/_next/static/chunks/webpack-6caf1cd8559327c9.js\" defer=\"\"></script><script src=\"/_next/static/chunks/framework-660abf7322751ffd.js\" defer=\"\"></script><script src=\"/_next/static/chunks/main-fb939c05790a6df3.js\" defer=\"\"></script><script src=\"/_next/static/chunks/pages/_app-0a0020ddd67f79cf.js\" defer=\"\"></script><script src=\"/_next/static/chunks/pages/_error-03529f2c21436739.js\" defer=\"\"></script><script src=\"/_next/static/zQgnelpe2z0ts-dbnrLzn/_buildManifest.js\" defer=\"\"></script><script src=\"/_next/static/zQgnelpe2z0ts-dbnrLzn/_ssgManifest.js\" defer=\"\"></script></head><body><div id=\"__next\"><div style=\"font-family:system-ui,&quot;Segoe UI&quot;,Roboto,Helvetica,Arial,sans-serif,&quot;Apple Color Emoji&quot;,&quot;Segoe UI Emoji&quot;;height:100vh;text-align:center;display:flex;flex-direction:column;align-items:center;justify-content:center\"><div style=\"line-height:48px\"><style>body{color:#000;background:#fff;margin:0}.next-error-h1{border-right:1px solid rgba(0,0,0,.3)}@media (prefers-color-scheme:dark){body{color:#fff;background:#000}.next-error-h1{border-right:1px solid rgba(255,255,255,.3)}}</style><h1 class=\"next-error-h1\" style=\"display:inline-block;margin:0 20px 0 0;padding-right:23px;font-size:24px;font-weight:500;vertical-align:top\">500</h1><div style=\"display:inline-block\"><h2 style=\"font-size:14px;font-weight:400;line-height:28px\">Internal Server Error<!-- -->.</h2></div></div></div></div><script id=\"__NEXT_DATA__\" type=\"application/json\">{\"props\":{\"pageProps\":{\"statusCode\":500}},\"page\":\"/_error\",\"query\":{},\"buildId\":\"zQgnelpe2z0ts-dbnrLzn\",\"nextExport\":true,\"isFallback\":false,\"gip\":true,\"scriptLoader\":[]}</script></body></html>"}
