+++ .wrangler/state/v3/r2/walkable-blog-cache/blobs/d90c78f8076cc1aac73bcc2af48a9b7b86be38c25f345709f9de5b54f1e18f7f000001985fbb6658
@@ -1,0 +1,1 @@
+{"type":"route","meta":{"status":200,"headers":{"cache-control":"public, max-age=0, must-revalidate","content-type":"text/plain","x-next-cache-tags":"_N_T_/layout,_N_T_/robots.txt/layout,_N_T_/robots.txt/route,_N_T_/robots.txt"}},"body":"User-Agent: *\nAllow: /\nDisallow: /private/\n\nSitemap: https://nim-fawn.vercel.app/sitemap.xml\n"}
